+++ tests/Element/HasUsemapTest.php
@@ -6,11 +6,13 @@ namespace UIAwesome\Html\Attribute\Tests\Element;
 
 use PHPUnit\Framework\Attributes\{DataProviderExternal, Group};
 use PHPUnit\Framework\TestCase;
+use Stringable;
 use UIAwesome\Html\Attribute\Element\HasUsemap;
 use UIAwesome\Html\Attribute\Tests\Support\Provider\Element\UsemapProvider;
 use UIAwesome\Html\Attribute\Values\ElementAttribute;
 use UIAwesome\Html\Helper\Attributes;
 use UIAwesome\Html\Mixin\HasAttributes;
+use UnitEnum;
 
 /**
  * Unit tests for the {@see HasUsemap} trait managing the `usemap` HTML attribute.
@@ -62,9 +64,9 @@ final class HasUsemapTest extends TestCase
      */
     #[DataProviderExternal(UsemapProvider::class, 'values')]
     public function testSetUsemapAttributeValue(
-        string|null $usemap,
+        string|Stringable|UnitEnum|null $usemap,
         array $attributes,
-        string $expectedValue,
+        string|Stringable|UnitEnum $expectedValue,
         string $expectedRenderAttribute,
         string $message,
     ): void {