+++ tests/Support/Stub/Values/Unit.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+namespace UIAwesome\Html\Attribute\Tests\Support\Stub\Values;
+
+/**
+ * Stub enum for tests.
+ *
+ * Provides deterministic values required by the test suite.
+ *
+ * @copyright Copyright (C) 2026 Terabytesoftw.
+ * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
+ */
+enum Unit
+{
+    /**
+     * Type representing a value.
+     */
+    case value;
+}